+++ include/swift/Localization/LocalizationFormat.h
