+++ lib/CMakeLists.txt
@@ -27,6 +27,7 @@ add_subdirectory(IDE)
 add_subdirectory(Immediate)
 add_subdirectory(IRGen)
 add_subdirectory(LLVMPasses)
+add_subdirectory(Localization)
 add_subdirectory(Markup)
 add_subdirectory(Migrator)
 add_subdirectory(Option)
