+++ lib/Localization/CMakeLists.txt
@@ -1,0 +1,6 @@
+
+add_swift_host_library(swiftLocalization STATIC
+  LocalizationFormat.cpp
+
+  LLVM_LINK_COMPONENTS
+    support)
